% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_powprod.R
\name{matched_powprod}
\alias{matched_powprod}
\title{Do a batch of power-products on a data frame}
\usage{
matched_powprod(df_weights, df_target)
}
\arguments{
\item{df_weights}{power weights (exponents)}

\item{df_target}{target data frame}
}
\value{
data frame with power-product results
}
\description{
Do a batch of power-products on a data frame
}
\examples{
df_weights <-
  tribble(
    ~out_name, ~X1, ~X2,
         "Y1",   1,   1,
         "Y2",   1,  -1
  )

df_target <-
 tibble(
   X1 = c( 1, 1, 1, 1),
   X2 = c( 2, 3, 4, 5)
 )

matched_powprod(df_weights, df_target)

bind_cols(
  df_target,
  matched_powprod(df_weights, df_target)
)
}
