% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_cdot.R
\name{named_cdot}
\alias{named_cdot}
\title{Do a single circle-dot on a data frame}
\usage{
named_cdot(weights, df_given)
}
\arguments{
\item{weights}{inner product weights; vector}

\item{df_target}{target data; data frame}
}
\value{
inner product results; vector
}
\description{
Do a single circle-dot on a data frame
}
\examples{
weights <- c("X1" = +0.5, "X2" = -0.5)

df_target <-
  tribble(
    ~X1, ~X2,
      4,   2,
      8,   6
  )

named_cdot(weights, df_target)
}
