% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_inner.R
\name{named_inner}
\alias{named_inner}
\title{Do a single inner product on a data frame}
\usage{
named_inner(weights, df_given)
}
\arguments{
\item{weights}{inner product weights; vector}

\item{df_target}{target data; data frame}
}
\value{
inner product results; vector
}
\description{
Do a single inner product on a data frame
}
\examples{
weights <- c(
  "X1" = 0.5,
  "X2" = 0,
  "X3" = -0.5
)

df_target <-
 tibble(
   X1 = c( 0, 1, 2, 3),
   X2 = c( 0, 0, 1, 1),
   X3 = c( 5, 6, 7, 8)
 )

named_inner(weights, df_target)
}
