% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_powprod.R
\name{named_powprod}
\alias{named_powprod}
\title{Do a single power-product on a data frame}
\usage{
named_powprod(weights, df_given)
}
\arguments{
\item{weights}{power weights (exponents); vector}

\item{df_target}{target data; data frame}
}
\value{
inner product results; vector
}
\description{
Do a single power-product on a data frame
}
\examples{
weights <- c(
  "X1" = +1,
  "X2" = -1,
)

df_target <-
 tibble(
   X1 = c( 1, 1, 1.0),
   X2 = c( 2, 3, 0.5),
 )

named_inner(weights, df_target)
}
