% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_nondim.R
\name{pi_nondim}
\alias{pi_nondim}
\title{Compute a unique non-dimensionalizing factor}
\usage{
pi_nondim(u_dim, df_dim)
}
\arguments{
\item{u_dim}{target dimensions; named vector, names must be in df_dim %>% pull(dim)}

\item{df_dim}{dimension matrix, must contain `dim` character column}
}
\value{
weights for non-dimensionalizing factor; named vector
}
\description{
Compute a unique non-dimensionalizing factor
}
\examples{
## Construct example
df_dim <-
  tribble(
    ~rho, ~U, ~d, ~mu, ~dim,
       1,  0,  0,   1,  "M",
      -3,  1,  1,  -1,  "L",
       0, -1,  0,  -1,  "T"
  )

u_dim <- c(
  "M" = 1, "L" = 1, "T" = -2
)

w_dim <- pi_nondim(u_dim, df_dim)

## Check proper dimensions
named_inner(w_dim, df_dim)

## Check orthogonality
df_proj <- pi_projection(df_dim)
named_inner(w_dim, df_proj)
}
