% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_projection.R
\name{pi_projection}
\alias{pi_projection}
\title{Generate a projection to the pi-subspace}
\usage{
pi_projection(df_dim)
}
\arguments{
\item{df_dim}{dimension matrix}
}
\value{
data frame for matched_inner()
}
\description{
Generate a projection to the pi-subspace
}
\examples{
## Basic example
df_dim <-
  tribble(
    ~rho, ~U, ~d, ~mu, ~dim,
       1,  0,  0,   1,  "M",
      -3,  1,  1,  -1,  "L",
       0, -1,  0,  -1,  "T"
  )

df_pi_proj <-
  pi_projection(df_dim)

df_target <-
  tribble(
    ~rho, ~U, ~d, ~mu,
     1.2, 10,  5, 1e-5,
     1.2, 20,  5, 1e-5,
     1.2, 10, 10, 1e-5,
     1.2, 20, 10, 1e-5
  )

matched_inner(df_pi_proj, df_target)

## Find basis orthogonal to given direction
df_dim2 <-
  tribble(
    ~a, ~b, ~c, ~d, ~e, ~dim,
     1,  1,  1,  1,  1,  "L",
     1, -1,  0,  0,  0,  ""
  )

df_pi_proj2 <-
  pi_projection(df_dim2)
}
