% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{Transpose a data frame}
\usage{
transpose(df_given)
}
\arguments{
\item{df_given}{input data frame with out_name}
}
\value{
transposed data frame
}
\description{
Transpose a data frame
}
\examples{
df_given <-
  tribble(
    ~out_name, ~X1, ~X2, ~X3,
         "W1",   1,   1,   0,
         "W2",  -1,   0,   1
  )

transpose(df_given)
}
