% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_projection.R
\name{pi_projection}
\alias{pi_projection}
\title{Generate a projection to the pi-subspace}
\usage{
pi_projection(df_dim)
}
\arguments{
\item{df_dim}{dimension matrix}
}
\value{
data frame for matched_inner()
}
\description{
Generate a projection to the pi-subspace
}
\examples{
df_dim <-
  tribble(
    ~rho, ~U, ~d, ~mu,
       1,  0,  0,   1,
      -3,  1,  1,  -1,
       0, -1,  0,  -1
  )

df_pi_proj <-
  pi_projection(df_dim)

df_target <-
  tribble(
    ~rho, ~U, ~d, ~mu,
     1.2, 10,  5, 1e-5,
     1.2, 20,  5, 1e-5,
     1.2, 10, 10, 1e-5,
     1.2, 20, 10, 1e-5
  )

matched_inner(df_pi_proj, df_target)
}
